server4/requirements.txt
Flask==2.1.0
flask-socketio==5.3.0
eventlet==0.33.3
annotated-types==0.7.0
anyio==4.4.0
asyncio==3.4.3
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.113.0
h11==0.14.0
idna==3.8
motor==3.5.1
pydantic==2.9.0
pydantic-core==2.23.2
pymongo==4.8.0
sniffio==1.3.1
starlette==0.38.4
typing-extensions==4.12.2
uvicorn==0.30.6
pytz
mysql-connector-python
flask-cors
Flask-SocketIO==5.3.0
Werkzeug==2.0.3
cryptography>=3.4.7